improvement: move module Hilt to di package
